ML/agents/langchain/simple/requirements.txt
langchain[openai]
python-dotenv
